refactor(seo): convert /characters to server component and fix ReDoS
characters/page.tsx を server component に分離し、フィルタ UI は
CharactersListClient へ移設。フィルタ状態を ?q= / ?properties= /
?skills= の URL クエリ駆動に変え、初期 HTML に通常キャラ全件が
含まれるよう SSR を強化する。

合わせて serchFilter.ts を削除。検索ロジックを RegExp(searchTerm,
"i") から normalize+includes に置き換えており、ユーザー入力を直接
RegExp に渡していた ReDoS リスクを根本除去する。
